Outside Sealed Gate
This is a fairly good save file since this is before entering the Esper World with every Esper able to be obtained and the opportunity to play with Mog as early as possible (whom has all but one dance). Gau has some Rages yet not that many so I may be uploading another one where he has more than the ones he has.